+++ src/QueryBuilder.ts
@@ -2,6 +2,7 @@ import {ClickHouse} from 'clickhouse';
 import {Query} from './Query';
 import {Logger} from 'winston';
 import {InsertQuery} from './InsertQuery';
+import {DeleteQuery} from './DeleteQuery';
 
 
 export class QueryBuilder {
@@ -19,5 +20,9 @@ export class QueryBuilder {
 
     public insert(): InsertQuery {
         return new InsertQuery(this.connection, this.logger);
+    }
+
+    public delete(): DeleteQuery {
+        return new DeleteQuery(this.connection, this.logger);
     }
 }