+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## v1.5.0 - 2022-12-22
+
+### Added
+- Added support for `DELETE` queries + unit tests
+- Added example to documentation for `DELETE` queries
+
 ## v1.4.0 - 2022-12-18
 
 ### Added